Clash spec path fix (#176)
* Remove no longer supported header from Clash Detection API calls

* Fix clash spec path issue
