+++ src/frontend-samples/clash-review-sample/ClashDetectionJsonData.ts
