+++ src/frontend-samples/clash-review-sample/ClashTable.tsx
